mtag_data: cache DB results
In the original code, we didn't cache the results because say you're in a
loop then the same expression could have multiple values.  Or maybe it
wasn't turned no because the cache wasn't cleared out at the end of a
function leading to a use after free.

So far, I haven't noticed any real speedup from caching the data...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>