test_xen: add script to build xen
Similarly to the kernel, add a script that analyzes Xen. This script
uses one-line-scan to wrap smatch in the build process. The script is
taken from the test_kernel script.

Different version of Xen might be built with different compilers.
Therefore, we allow to specify the environment variable
SMATCH_ONE_LINE_SCAN_ARSG, which can be used to forward command line
parameters to one-line-scan, for example compiler prefix or suffix.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>