replace nbr_users() & multi_users() by one_use()
During simplification, we're only interested to know if a pseudo
is used only once or more than once. This can be checked quicker
than getting the exact number of users.

So, replace the last call to nbr_users() and the calls to multi_users()
by a call to one_use() which has a slightly clearer name.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>