user_data: fix validation test for returned struct members
The bug here can be seen from looking at the validation test.  The problem
is that p->y is both a parameter and a return and we were only recording
that it was a parameter.

This needs to be split out so it records that both the parameter
information and the return information.

Reported-by: Norbert Manthey <nmanthey@amazon.de>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>