switch return order in simplify_mask_or_and()
This function used an early return when the test guarding
the simplification failed but this make impossible to add
other ones simplifications.

In preparation for incoming additional simplifications,
negate the test and switch the returns so that the early return
is now a catch-all return when no simplifications can apply.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>