add a flag for volatile memops
When simplifying memops, it's needed to check if the memops
is a volatile access or not. This is currently done by checking
insn->type->ctype.modifiers & MOD_VOLATILE which is rather long
but also incorrect for bitfields.

Prepare to fix this by adding a flag to struct instruction directly
telling if the access is volatile.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>