simplify OP((x | C), K) when (C & M) == M
In an expression like OP((x | C), K), if the effective mask (M)
corresponding to OP(_, K) is equal to the combined mask (C & M),
then the OR operation is unneeded and can be replaced by M itself,
giving: OP(M, K).
In mathematical terms:
	0) ((x | C) & M) = ((x & M) | (C & M))
	1) (C & M) = M
	2) ((x & M) | (C & M)) = ((x & M) | M) = M
and so OP((x | C), K) -> OP(M, K).

For example, code like:
	unsigned int foo(int x)
	{
		return (x | 7) & 2;
	}

is now simplified into:
	foo:
		ret.32      $2

which previously was not optimized
	foo:
		or.32       %r2 <- %arg1, $7
		and.32      %r3 <- %r2, $2
		ret.32      %r3

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>