files: check for schema files relative to the bin_dir
One thing that we want to have is smatch_data/ directories inside the
project.  Most of the time the data can built automatically so that's
fine.

But we don't want to have to copy all the smatch_data/db/ stuff there,
because that doesn't change from project to project.  So with this patch
we check for the db file by searching from the bin_dir.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>