unify simplify_lsr_or() & simplify_and_or_mask()
Doing a LSR(X, S) will drop the S right bits.
So any simplification that can be made when using an AND clearing
the right S bits can also be used on LSR (as if its first operand
would first be implicitly be ANDed with such a mask).

So, in order to not duplicate complex simplifications involving
ANDs & ORs masks, merge these both function in a single one,
using the mask corresponding to the operation.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>