simplify ((x & M) >> S) when (M >> S) == 0
The instructions LSR(AND(x, M), S) are already simplified into
AND(LSR(x, S), (M >> S)) but only if AND(x, M) has a single user.

However, if (M >> S) == 0, they can always directly be simplified to 0.
For example code like:
	unsigned foo(unsigned x)
	{
		unsigned t = (x & 0x00000fff);
		return (t >> 12) & t;
	}

is now simplified into:
	foo:
		ret.32      $0

while previously it was:
	foo:
		and.32      %r2 <- %arg1, $0xfff
		lsr.32      %r4 <- %r2, $12
		and.32      %r6 <- %r4, %r2
		ret.32      %r6

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>