simplify ((x & M) >> S) when (M >> S) == (-1 >> S)
The instructions LSR(AND(x, M), S) are already simplified into
AND(LSR(x, S), (M >> S)) but only if AND(x, M) has a single user.

However, if (M >> S) == (-1 >> S), the AND part is redundant and the
whole can always directly be simplified into LSR(x, S).

For example, code like:
	unsigned foo(unsigned x)
	{
		unsigned t = (x & 0xfffff000);
		return ((t >> 12) ^ (x >> 12)) & t;
	}

is now optimized into:
	foo:
		ret.32      $0

because (t >> 12) is simplified into (x >> 12).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>