add testcases for {LSR,SHL}(AND(x, M), S) with shared AND(x, M)
The pattern LSR(AND(x, M), S) is already generically simplified into
((x >> S) & (M >> S)) but only if the sub-expression AND(x, M) is not
shared with some other expressions because the simplification modify it.

But for some special cases the expression can be simplified even if
the sub-expression is shared because the simplification doesn't need
to modify this AND(x, M) part.

Add the testcases for LSR and the incoming SHL.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>