doc: automatically insert the blank line for lists
ReST requires a blank line to start a list.

This is a bit annoying for the internal documentation because
this makes, IMO, things less readable in the code.

Tweak the documentation extraction tool to automatically insert
these blank lines (between a line ending with ':' and the next one
if not empty).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>