doc: simplify the creation of the viewlist
Each bloc-comment is entered in the view list as a single multiline
plus a final blank line. This is more complicated than needed and it's
useless to first concatenate these lines to split them directly after.
Also it makes harder to correctly track the line numbers.

Fix this by adding the lines one by one, each with their own lineno.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>