prepare simplification of MASK(SHIFT(a | b, S), M)
Currently the simplification of SHIFT(a | b, S) is done by
calling simplify_mask_or() with an effective mask corresponding
to the shift instruction.

Prepare the simplification of more complex patterns, involving
a combination of masks, by using a small helper, simplify_mask_shift_or(),
for doing the call to simplify_mask_or() with the appropriate mask.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>