simplify TRUNC(SHIFT(a | b, S), N)
The simplification of TRUNC(SHIFT(a | b, S), N) can be done by
combining the effective mask corresponding to TRUNC(_, N) with
the one corresponding to SHIFT(_, S).

This allows to also simplify signed bitfields. For example, code like:
	struct s {
		  signed int  :2;
		  signed int f:3;
	};
	int bfs(struct s s, int a)
	{
		s.f = a;
		return s.f;
	}

is now simplified into the minimal:
	bfs:
		trunc.3     %r4 <- (32) %arg2
		sext.32     %r11 <- (3) %r4
		ret.32      %r11

The simplification is done by calling simplify_mask_shift() with
the mask corresponding to TRUNC(_, N).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>