check_check_deref: silence some false positives
There are two things here.  We ignore some checks where we know that the
variable is non-NULL already.  I don't think this matters except that it
means we save fewer states.  We already ignored those later.  (Hopefully...
I've been running this code some time, and can't remember writing it).

The other thing is that if we've already checked for NULL then set the
non-NULL side to &ok.  Normally we wouldn't set the true side because if
it's no state that's fine.  But the problem is when we check for null twice
in a row and on the second time then the state is &null on both sides.
(I'm not sure why implications are failing in this situation...)  Anyway on
the NULL side, we reset the variable to an unknown value.

I guess what I'm saying is that this change is obviously harmless, and I
know from testing that it's required.  But in retrospect, I wish I had
investigated some more to find out why it's required instead of being
dealt with by the implication code.  One possible explanation was that
there were some impossible states and the implication code doesn't deal
well with impossible states which is a reasonable explanation, but if I
remember correctly, it's not the correct explanation.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>