container_of: fix a bug in get_shared_cnt()
The get_shared_cnt() is supposed to take a function pointer such as
foo->bar->func() and a parameter foo->bar->x and say that they have
"foo->bar" shared in common.  The bug is that it was taking a function
like "foo_bar()" and a parameter like "foo_param" and saying that they
have "foo_" in common.  When actually they're not related at all except
by chance that their names start with the same letters.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>