mtag_data: remove duplicate mtag_data and random small cleanups
The one behavior change here is that we call "delete from mtag_data" to
delete the old data.  We weren't supposed to have duplicate lines.

The change to save_mtag_data() is purely cosmetic.  When we preserve the
mtag_data, I only want to preserve DATA_VALUE information because that's
the only one that does the weird thing with saving pointers in the DB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>