fix linearization of unreachable switch (with reachable label).
An unreachable/inactive switch statement is currently not
linearized. That's nice because it avoids to create useless
instructions.

However, the body of the statement can contain a label which
can be reachable. If so, the resulting IR will contain a branch
to an unexisting BB. Bad.

For example, code like:
	int foo(int a)
	{
		goto label;
		switch(a) {
		default:
	label:
			break;
		}
		return 0;
	}

(which is just a complicated way to write: int foo(int a) { return 0; })

is linearized as:
	foo:
		br          .L1

Fix this by linearizing the statement even if not active.

Note: it seems that none of the other statements are discarded
      if inactive. Good. OTOH, statement expressions can also
      contains (reachable) labels and thus would need the same
      fix (which will need much more work).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>