trivial-phi: directly return the unique value
In trivial_phi(), the fact that the phi-node is trivial or not
is returned as an int and, if trivial, the unique value is
returned via the pointer given as first argument.

But these two results can easily be combined in a single one
by returning the unique value if trivial and NULL otherwise.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>