has-attr: add support for __has_attribute()
Sparse has support for a subset of GCC's large collection of
attributes. It's not easy to know which versions support this
or that attribute. However, since GCC5 there is a good solution
to this problem: the magic macro __has_attribute(<name>)
which evaluates to 1 if <name> is an attribute known to the
compiler and 0 otherwise.

Add support for this __has_attribute() macro by extending the
already existing support for __has_builtin().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>