mtag: add string information
This code isn't totally perfect, but if I don't commit it now, then I'm
going to just end up forgetting what's involved...

Basically we store a string in mtag_data and we have an mtag which points
to it.  This is only used to calculate strlen() for known strings, but
eventually it will be used more often.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>