return nothing only in void functions
Currently, the code for the return is only generated if the
effectively return a type or a value with a size greater than 0.

But this mean that a non-void function with an error in its return
expression is considered as a void function for what the generated
IR is concerned, making things incoherent.

Fix this by using the declared type instead of the type of the
return expression.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>