fix linearization of non-constant switch-cases
The linearization of switches & cases makes the assumption
that the expressions for the cases are constants (EXPR_VALUE).
So, the corresponding values are dereferenced without checks.

However, if the code uses a non-constant case, this dereference
produces a random value, probably one corresponding to some
pointers belonging to the real type of the expression.

Fix this by checking during linearization the constness of the
expression and ignore the non-constant ones.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>