db: fix a type bug
The issue here was there was a function which returned u16 but we had
"ret = -1;".  It turns out the signedness bug doesn't complain if you
return -1 in a u16 function because I guess that's idiomatic.

Anyway, in the DB the type was wrong so it recorded that we were returning
-1 instead of U16_MAX.  So the caller parsed it incorrectly and got
confused.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>