test: make test Waddress-space-strict succeed on 32-bit
The test Waddress-space-strict made assumptions about the
relative size of integers & pointers. Since this test was
crafted on a 64-bit machine, the test was running fine for
LP64 but failed on a 32-bit machine (or anything using IP32,
like using the -m32 option).

However, since the test is about conversion of address-spaces,
using integers of different size adds no value, and indeed
brings problems.

Fix this by limiting the conversions to a single integer type,
the one with the same size as pointers on ILP32 & LP64: long.

Reported-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>