array: store possible array values for simple arrays
The concept is simple enough.  We want smatch to be able to handle
__smatch_implied(array[idx]);.  This is a very limited first implementation
because it only handles file scoped arrays.  I will add more later.

Also the table in the DB is called "sink_info" because I hope that it will
eventually be more general purpose than just handling arrays.  Hopefully,
it will hand lists as well later.  Where we store information about what
we're putting into a list and what we pull out from the list.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>