extra: don't reset impossible states after a condition
Say you have "x" set to an empty state, then would do "if (x != 3)" then it
was setting the true and false states.  We don't want that, we want
empty to remain empty.

The issue I think is that we started using get_real_absolute_rl() to get
the rl.  That function retuns the whole rl for empty states.  We don't have
a way to return an empty rl, only a NULL rl otherwise we'd maybe look at
changing get_real_absolute_rl().  This might be the only situation where it
matters so maybe this is fine.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>