type_val: don't double record globals
I changed smatch_flow.c a bit with regards to how global assignments are
handled.  It doesn't affect the output, but in the DB we record globals
as being set globally and also being set inside the next function.  It's
confusing but harmless unless you are trying to debug the code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>