nospec: don't fill the DB up with garbage
The problem is that everything was getting marked as nospec.  I don't
really know why this is...  The get_user_stree() was only supposed to
return user data so this should have worked from the beginning.

Anyway, I took the big hammer approach to fixing it and add a second check
to make sure that everything we save in the DB is also recorded as user
data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>