param_limit: make mtag limits more generic
Say a function is only called with a known mtag parameter.  Then it has
a check for NULL "if (!p) return -EINVAL;" that is an impossible path and
we CULL it.  Fine.  But then the success path is limited to a specific
mtag.

So then we add a new caller, which passes a different mtag.  Then smatch is
like this mtag is impossible on every return path...  And it creates a
bunch of problems.  So what this does is we replace mtag limits with
"valid pointer" instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>