extra/type: fix type of return PARAM_ADD/SET
Imagine we have a function "void frob(char *p);" and it sets *p to an
unknown signed char value.  Then we have a caller "frob(&u8_variabl);".
Before we would take -128-127 from the DB and try to put it into a u8,
but -128 is less than a u8 so it would get confused and set the result to
0-127.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>