param_set: handle when the stack variable was modified better
Say you have a function like strcmp() which takes a parameter s1 does
"s1++;".  We need to save that "s1" was modified, because we want
param_was_set() to work.  We don't want to save that *s1 was set, because
it actually wasn't.  And then generally if the stack variable changes, we
don't want to put that *s1 type stuff was changed (because it wasn't) so
we need to insert "s1" into the &set_list.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>