enum: keep enumerators as int if they fit
In Standard C, enumerators have type 'int' and an unspecified
base type. OTOH, GCC (and thus sparse) llows any integer type
(and sparse also more or less allows bitwise types).
After the enum's decalration is parsed, the enumerators are
converted to the underlying type.

Also, GCC (and thus sparse) uses an unsigned type unless one of
the enumerators have a negative value.

This is a problem, though, because when comparing simple integers
with simple enumerators like:
	enum e { OK, ONE, TWO };
the integers will unexpectedly be promoted to unsigned.
GCC avoid these promotions by not converting the enumerators that
fit in an int.

For GCC compatibility, do the same: do not convert enumerators
that fit in an int.

Note: this is somehow hackish but without this some enum usages
      in the kernel give useless warnings with sparse.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>