user_data: integers can't be dereferenced
There are places in the kernel where we use unsigned longs instead of
pointers.  Smatch tries to no lose track of information about the members
of those struct pointers.  But say we have an int, then it can't be a
struct pointer so we don't have to know about the struct members.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>