__attribute__((fallthrough)) can't simply be ignored
Currently, sparse has the attribute 'fallthrough' in its list of
known-but-ignored attributes (like almost every GCC's attributes).

But this attribute is a statement attribute, something which is
currently not supported (it's interpreted as the attribute of an empty
declaration which doesn't play well with -Wdeclaration-after-statement).

Fix this by stopping to consider this attribute as known. This will
allow __has_attribute(fallthrough) to correctly play its role.

Note: a more complete solution will need to parse this statement
      attribute and maybe be able to make the distinction between
      statement attributes, label attributes (also currently ignored),
      and type, function & variable attributes.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>