nul_terminator: track when strings have a NUL terminator
This code is very very incomplete and I haven't really tested it.  But I
will expand it later.

It's supposed to track if a string is known to be NUL terminated or known
to be not NUL terminated.  My guess is that it's only manual annotation of
places which are dangerous which will mark the string as non-terminated.
So like copy_from_user() will mark the string as non-NUL terminated.

Anyway, I have some places where this will be useful right off the bat so
I am going to commit it as-is for now.  We'll see how it evolves in the
future.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>