pre-process: print macros containing # and ## correctly
The dump_macro() function fails to correctly output the definitions of
macros that contain the string operator '#', the concatenation operator
'##' and any macro parameter in the definition token list. For example,
the following macros:

    #define STRING(x) #x
    #define CONCAT(x,y) x ## y

are output like so:

    #define STRING(x) unhandled token type '21'
    #define CONCAT(x, y) unhandled token type '22'  unhandled token type '23'  unhandled token type '22'

Add the code necessary to handle those token types to the dump_macros()
function and add the above macros to the 'dump-macros.c' test file.

Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>